+++ position/src/lib.rs
@@ -6,4 +6,5 @@
 pub mod bitboard;
 pub mod board;
 pub mod position;
+pub mod move_generation;
 pub mod prelude;
